Add a button/popup to enable registering cmd.exe
Registering cmd.exe directly through IOTCoreDefaultApp would make it
easier to use the command line processor page. How it works: if the user
is not yet authorized to run cmd.exe, a button is displayed. Clicking
the button will popup a dialog to let a user enter the admin
username/password on the device. A command is then sent to the device
using the portal API to register cmd.exe and success/failure is reported
back to the user.