MSFT: 8554547  This change adds passphrase conversion support on the IotOnboarding producer.
The AllJoyn Onboarding protocol expects that the passphrase or network key is converted to a
HEXized Ascii String (e.g. “Apple123” converts to "4170706C65313233").  On the AllJoyn Producer
side however, the HEXized string must be converted back to a passphrase to make the WinRT call
with the correct passphrase.  This code converts the string back before onboarding the device
to the desired network.