Fix for Bug in IOT Default App for soft AP
IOT Default App (Bertha) does not show Soft-AP IP Address when the wired
internet connection is removed (e.g. when only the Soft-AP is enabled
and no Ethernet is plugged in).
When the device has no wired or wired connections, the
HostName.IPInformation.NetworkAdapter API does not return a network
adapter when enumerating IP addresses, even when an IP address is
associated with soft-AP. This result on soft-AP not displayed in Bertha
device info page.
To resolve this, introduce new logic to check if a network adapter has
been detected. If not, a virtual adapter is assumed and associated with
the the IP address found.